+++ backend/src/main/java/br/edu/ifac/napne360/Napne360BackendApplication.java
@@ -1,0 +1,13 @@
+package br.edu.ifac.napne360;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Napne360BackendApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(Napne360BackendApplication.class, args);
+	}
+
+}