+++ backend/src/test/java/br/edu/ifac/napne360/Napne360BackendApplicationTests.java
@@ -1,0 +1,13 @@
+package br.edu.ifac.napne360;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class Napne360BackendApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}